pyproject.toml
[project]
name = "sip-to-ai"
version = "0.1.0"
description = "Bidirectional audio bridge between SIP/PJSUA2 and AI realtime voice services"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "GPL-2.0"}
authors = [
    {name = "SIP-to-AI Team", email = "team@example.com"}
]
dependencies = [
    "websockets>=15.0.1",
    "numpy>=2.3.3",
    "soxr>=1.0.0",
    "structlog>=25.4.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "mypy>=1.18.2",
    "ruff>=0.13.2"
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"

[[tool.mypy.overrides]]
module = [
    "pjsua2.*",
    "soxr.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long - handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.2",
    "types-pyyaml>=6.0.12.20250915",
]
